feat(rust-project): refactor rust project in several components + share custom gitignore
