+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './rust/RustProject';
-export * from './typescript/TypescriptLibraryProject';
-export * from './typescript/TypescriptApplicationProject';
+export * from './git';
+export * from './rust';
+export * from './typescript';
