+++ src/typescript/index.ts
@@ -1,0 +1,2 @@
+export * from './TypescriptApplicationProject';
+export * from './TypescriptLibraryProject';
